requirement.txt
pandas
requests
beautifulsoup4
camelot-py[cv]
openpyxl